==> barlow_reanalysis/03_Barlow_BetaDiversity.qmd <==
---
title: "Comparing Samples with Beta Diversity" 
author: "Augustus Pendleton"
date: "`r format(Sys.time(), '%d %B, %Y')`"
editor_options: 
  chunk_output_type: console
knitr: 
  opts_chunk: 
    dpi: 300
    dev: png
    dev.args: { bg: "transparent" }
---
# Load Data and Packages

```{r load-packages}

pacman::p_load(phyloseq, ggdendro, dendextend, vegan, patchwork, GUniFrac, tidyverse, ggpubr)

# Make sure we're using the MarschmiLab version, which allows us to incorporate absolute abundance!
packageDescription("GUniFrac")

load("root_physeq.RData")

source("../code/subset_dist_matrix.R")
source("../code/plotting.R")

```

# Explore samples via ordination

Conflicted with the best way to approach rarefaction and absolute values here. I'm less confident than I was with Ontario that I'm reaching equal saturation across my samples, and I don't want my beta diversity to be confounded by richness due to sequencing depth. What I settled on:


The generate_rarefied_abs_tables returns these rarefied, absolute-count normalized tables in the form of a 3D-array, where dim 1 is each sample, dim 2 is each iteration of rarefaction, and dim 3 is each ASV

```{r, generating-abs-dist-objects}

#install.packages("abind")
#install.packages("extraDistr")
source("../code/generate_rarefied_abs_tables.R")

cell_counts_prep <- 
  root_physeq %>%
  sample_data %>%
  data.frame %>%
  select(sample, Sample_Sum)

cell_counts <- cell_counts_prep$Sample_Sum

names(cell_counts) = cell_counts_prep$sample

abs_tables <- generate_rarefied_abs_tables(physeq = root_physeq, # Physeq with non-normalized counts (do not convert to rel abundance!)
                              iterations = 100, # Number of rarefy steps
                              rare.depth = min(sample_sums(root_physeq)), # Depth to rarefy to
                              abs.counts = cell_counts, # NAMED vector of cell counts for each sample
                              seed = 314,
                              return_rel = FALSE) # Random seed for reproducibility

dims <- dim(abs_tables)

abs_bray_dists <- 
  map(c(1:dims[2]), 
    \(x)vegdist(abs_tables[,x,], method = "bray"))

abs_bray_dist <- Reduce(`+`, abs_bray_dists) / length(abs_bray_dists)


# abs_gun_dists <-
#   map(c(1:dims[2]),
#     \(x)GUniFrac(otu.tab = abs_tables[,x,],
#                  tree = phy_tree(root_physeq),
#                  alpha = seq(from = 0, to = 1, by = 0.1),
#                  normalize_counts = FALSE,
#                  verbose = FALSE))
# 
# save(abs_gun_dists, file = "abs_gun_dists.RData")

load("abs_gun_dists.RData")

gun_names <- c("d_UW","d_0","d_0.1","d_0.2","d_0.3","d_0.4","d_0.5","d_0.6","d_0.7","d_0.8","d_0.9","d_1")

names(gun_names) <- gun_names

abs_unifracs <- 
  map(gun_names, \(x){
  dist_objs <- map(abs_gun_dists,
                   \(y)as.dist(y$unifracs[,,x]))
  
  Reduce(`+`, dist_objs) / length(dist_objs)
})

```

```{r, generating-rel-dist-objects}

rel_counts <- rep(1, 93)

names(rel_counts) <- names(cell_counts)

rel_tables <- generate_rarefied_abs_tables(physeq = root_physeq, # Physeq with non-normalized counts (do not convert to rel abundance!)
                              iterations = 100, # Number of rarefy steps
                              rare.depth = min(sample_sums(root_physeq)), # Depth to rarefy to
                              abs.counts = rel_counts, # NAMED vector of cell counts for each sample
                              seed = 314,
                              return_rel = TRUE) # Random seed for reproducibility


rel_bray_dists <- 
  map(c(1:dims[2]), 
    \(x)vegdist(rel_tables[,x,], method = "bray"))

rel_bray_dist <- Reduce(`+`, rel_bray_dists) / length(rel_bray_dists)


# rel_gun_dists <-
#   map(c(1:dims[2]),
#     \(x)GUniFrac(otu.tab = rel_tables[,x,],
#                  tree = phy_tree(root_physeq),
#                  alpha = seq(from = 0, to = 1, by = 0.1),
#                  normalize_counts = FALSE,
#                  verbose = FALSE))
# 
# save(rel_gun_dists, file = "rel_gun_dists.RData")

load("rel_gun_dists.RData")

gun_names <- c("d_UW","d_0","d_0.1","d_0.2","d_0.3","d_0.4","d_0.5","d_0.6","d_0.7","d_0.8","d_0.9","d_1")

names(gun_names) <- gun_names

rel_unifracs <- 
  map(gun_names, \(x){
  dist_objs <- map(rel_gun_dists,
                   \(y)as.dist(y$unifracs[,,x]))
  
  Reduce(`+`, dist_objs) / length(dist_objs)
})

```

Comparing dist values

```{r compare-dists, fig.width = 7.5, fig.height = 2.5}

bcr_au <- rel_bray_dist - abs_unifracs$d_1 

ur_au <- rel_unifracs$d_1 - abs_unifracs$d_1

bca_au <- abs_bray_dist - abs_unifracs$d_1

data.frame(comparison = rep(c("BCR - UA", "UR - UA", "BCA - UA"), each = 4278),
           diff = c(bcr_au, ur_au, bca_au)) %>%
  ggplot(aes(x = diff)) + 
  geom_histogram(bins = 100) + 
  geom_vline(xintercept = 0, linetype = 2, linewidth= 0.2) + 
  facet_wrap(~comparison)


data.frame(ua = rep(c(abs_unifracs$d_1), 3),
           other_dist = c(c(rel_bray_dist),c(rel_unifracs$d_1), c(abs_bray_dist)),
           dist_name = rep(c("BR", "UR", "BA"), each = 4278)) %>%
  mutate(dist_name = factor(dist_name, levels = c("BR","UR", "BA"),
                             labels = c(expression(italic(BC^{R})),
                                        expression(italic(U^{R})),
                                        expression(italic(BC^{A}))))) %>%
  ggplot(aes(x = ua, y = other_dist)) + 
  stat_density_2d(geom = "polygon", contour = TRUE,
                  aes(fill = after_stat(level)),
                  bins = 30) +
  scale_fill_distiller(palette = "Blues", direction = 1) + 
  facet_wrap(~dist_name, labeller = label_parsed) + 
  geom_abline(slope = 1, intercept = 0, linetype = 2, linewidth = 0.2) + 
  scale_x_continuous(breaks = c(0,0.5,1)) + 
  scale_y_continuous(breaks = c(0,0.5,1)) + 
  coord_fixed(xlim = c(0,1), ylim = c(0,1)) + 
  labs(x = expression(italic(U^{A})), y = "Other Distance", caption = "n = 4278") + 
  theme(legend.position = "none")

```

# Picking which axis to compare across

```{r choose-axis}

barlow_sam <- root_physeq %>%
  sample_data() %>%
  data.frame()

barlow_sam %>%
  ggplot(aes(fill = Host_diet, x = Sample_Sum)) + 
  geom_histogram(alpha = 0.5) +
  facet_wrap(~host_tissue_sampled) 


```

Okay, we're gonna compare diets (the main intervention), but just within stool. 

```{r permanovas_and_mantels, fig.width = 4, fig.height = 4}

mature_samples <- 
  barlow_sam %>%
  dplyr::filter(host_tissue_sampled == "Stool") %>%
  pull(sample)

sm_abs_bray_dist <- subset_dist_matrix(abs_bray_dist, mature_samples)
sm_rel_bray_dist <- subset_dist_matrix(rel_bray_dist, mature_samples)
sm_rel_unifracs <- map(rel_unifracs, subset_dist_matrix, keeps = mature_samples)
sm_abs_unifracs <- map(abs_unifracs, subset_dist_matrix, keeps = mature_samples)

mature_df <- 
  barlow_sam %>%
  dplyr::filter(host_tissue_sampled == "Stool")

abs_bray_perm <- adonis2(sm_abs_bray_dist~Host_diet, data = mature_df)
rel_bray_perm <- adonis2(sm_rel_bray_dist~Host_diet, data = mature_df)

rel_perms <- map(sm_rel_unifracs, \(x){
  adonis2(x~Host_diet, data = mature_df)
})

abs_perms <- map(sm_abs_unifracs, \(x){
  adonis2(x~Host_diet, data = mature_df)
})

rel_rs <- unlist(map(rel_perms, \(x)x$R2[1]))
abs_rs <- unlist(map(abs_perms, \(x)x$R2[1]))

data.frame(rs = c(rel_rs[-1], abs_rs[-1]),
           alpha = rep(seq(0, 1, by = 0.1), 2),
           metric = rep(c("Relative","Absolute"), each = 11)) %>%
  ggplot(aes(x = alpha, y = rs, color = metric)) + 
  annotate(geom = "segment", y = rel_bray_perm$R2[1], yend = rel_bray_perm$R2[1],
           x = 0, xend = 1, linetype = 2, color = "#196689", linewidth = 2) + 
  annotate(geom = "segment", y = abs_bray_perm$R2[1], yend = abs_bray_perm$R2[1],
           x = 0, xend = 1, linetype = 2, color = "#B6AA0D", linewidth = 2) +  
    annotate(geom = "segment", y = rel_rs[1], yend = rel_rs[1],
           x = 0, xend = 1, linetype = 2, color = "darkred", linewidth = 2) + 
  labs(x = "\u03b1", y = expression(italic(R^2))) + 
  geom_line(linewidth = 2) + 
  scale_x_continuous(expand = expansion(mult = c(0, 0.05)))+ 
  scale_color_manual(values = c("#F2C621", "#008FF8"))+ 
  theme(legend.position = "none",
        axis.title = element_text(size = 20)) + 
  coord_fixed(xlim = c(0,1), ylim = c(0,1))

rel_f <- unlist(map(rel_perms, \(x)x$F[1]))
abs_f <- unlist(map(abs_perms, \(x)x$F[1]))

data.frame(f = c(rel_f[-1], abs_f[-1]),
           alpha = rep(seq(0, 1, by = 0.1), 2),
           metric = rep(c("Relative","Absolute"), each = 11)) %>%
  ggplot(aes(x = alpha, y = f, color = metric)) + 
  annotate(geom = "segment", y = rel_bray_perm$F[1], yend = rel_bray_perm$F[1],
           x = 0, xend = 1, linetype = 2, color = "#196689", linewidth = 2) + 
  annotate(geom = "segment", y = abs_bray_perm$F[1], yend = abs_bray_perm$F[1],
           x = 0, xend = 1, linetype = 2, color = "#B6AA0D", linewidth = 2) +  
    annotate(geom = "segment", y = rel_f[1], yend = rel_f[1],
           x = 0, xend = 1, linetype = 2, color = "darkred", linewidth = 2) + 
  labs(x = "\u03b1", y = expression(italic(pseudo-F))) + 
  geom_line(linewidth = 2) + 
  scale_x_continuous(expand = expansion(mult = c(0, 0.05)))+ 
  scale_color_manual(values = c("#F2C621", "#008FF8"))+ 
  theme(legend.position = "none",
        axis.title = element_text(size = 20)) + 
  coord_fixed(ratio = 1/(max(rel_bray_perm$F[1], abs_bray_perm$F[1], rel_f,  abs_f) - min(rel_bray_perm$F[1], abs_bray_perm$F[1], rel_f,  abs_f)))

rel_p <- unlist(map(rel_perms, \(x)x$`Pr(>F)`[1]))
abs_p <- unlist(map(abs_perms, \(x)x$`Pr(>F)`[1]))

data.frame(f = c(rel_p[-1], abs_p[-1]),
           alpha = rep(seq(0, 1, by = 0.1), 2),
           metric = rep(c("Relative","Absolute"), each = 11)) %>%
  ggplot(aes(x = alpha, y = f, color = metric)) + 
  annotate(geom = "segment", y = rel_bray_perm$`Pr(>F)`, yend = rel_bray_perm$`Pr(>F)`[1],
           x = 0, xend = 1, linetype = 2, color = "#196689", linewidth = 2) + 
  annotate(geom = "segment", y = abs_bray_perm$`Pr(>F)`[1], yend = abs_bray_perm$`Pr(>F)`[1],
           x = 0, xend = 1, linetype = 2, color = "#B6AA0D", linewidth = 2) +  
    annotate(geom = "segment", y = rel_p[1], yend = rel_p[1],
           x = 0, xend = 1, linetype = 2, color = "darkred", linewidth = 2) + 
  labs(x = "\u03b1", y = expression(italic(p))) + 
  geom_line(linewidth = 2) + 
  scale_x_continuous(expand = expansion(mult = c(0, 0.05)))+ 
  scale_color_manual(values = c("#F2C621", "#008FF8"))+ 
  theme(legend.position = "none",
        axis.title = element_text(size = 20)) + 
  scale_y_log10(limits = c(0.001, 1)) + 
  theme(aspect.ratio = 1)

cell_dist <- mature_df %>%
  dplyr::select(Sample_Sum) %>% 
  dist()


dist_list <- list(Bray_Rel = sm_rel_bray_dist,
                  Bray_Abs = sm_abs_bray_dist) %>%
  append(sm_rel_unifracs) %>%
  append(sm_abs_unifracs)

mantels <- unlist(map(dist_list, \(x)vegan::mantel(cell_dist, x, permutations = 999)$statistic))


data.frame(mans = mantels[c(4:14, 16:26)],
           alpha = rep(seq(0, 1, by = 0.1), 2),
           metric = rep(c("Relative","Absolute"), each = 11)) %>%
  ggplot(aes(x = alpha, y = mans, color = metric)) + 
  annotate(geom = "segment", y = mantels[1], yend =  mantels[1],
           x = 0, xend = 1, linetype = 2, color = "#196689", linewidth = 2) + 
  annotate(geom = "segment", y =  mantels[2], yend =  mantels[2],
           x = 0, xend = 1, linetype = 2, color = "#B6AA0D", linewidth = 2) +  
    annotate(geom = "segment", y =  mantels[3], yend =  mantels[3],
           x = 0, xend = 1, linetype = 2, color = "darkred", linewidth = 2) + 
  labs(x = "\u03b1", y = "Mantel R") + 
  geom_line(linewidth = 2) + 
  scale_x_continuous(expand = expansion(mult = c(0, 0.05)))+ 
  scale_color_manual(values = c("#F2C621", "#008FF8"))+ 
  theme(legend.position = "none",
        axis.title = element_text(size = 20)) + 
  coord_fixed(xlim = c(0,1), ylim = c(0,1))
```

# Ordinations for Supplemental

```{r ordinations, fig.width = 4, fig.height=4}

mature_physeq <- subset_samples(root_physeq, sample %in% mature_samples)
plot_ordination(
  mature_physeq,
  ordination = ordinate(mature_physeq, method = "PCoA", distance = sm_abs_unifracs$d_1),
  color = "Host_diet"
) + 
  theme(legend.box.background = element_rect(color = NA),
        aspect.ratio = 1) + 
  scale_color_manual(values = c("#196689", "#B6AA0D", "darkred")) + 
  labs(color = "Host Diet", size = "Copy #/gram")+ 
  geom_point(aes(size = Sample_Sum), alpha = 0.6) + 
  scale_size_continuous(labels = scales::label_log())

plot_ordination(
  mature_physeq,
  ordination = ordinate(mature_physeq, method = "PCoA", distance = sm_rel_unifracs$d_1),
  color = "Host_diet"
)+ 
  theme(legend.box.background = element_rect(color = NA),
        aspect.ratio = 1) + 
  scale_color_manual(values = c("#196689", "#B6AA0D", "darkred")) + 
  labs(color = "Host Diet")
  stat_ellipse()

```

```{r ordinations-noleg-abs, fig.width = 4, fig.height=4}

mature_physeq <- subset_samples(root_physeq, sample %in% mature_samples)
plot_ordination(
  mature_physeq,
  ordination = ordinate(mature_physeq, method = "PCoA", distance = sm_abs_unifracs$d_1),
  color = "Host_diet"
) + 
  theme(legend.box.background = element_rect(color = NA),
        aspect.ratio = 1) + 
  scale_color_manual(values = c("#196689", "#B6AA0D", "darkred")) + 
  labs(color = "Host Diet") + 
  theme(legend.position = "none")+ 
  geom_point(aes(size = Sample_Sum), alpha = 0.6) + 
  scale_size_continuous(labels = scales::label_log())

```

```{r ordinations-noleg-rel, fig.width = 4, fig.height=4}

mature_physeq <- subset_samples(root_physeq, sample %in% mature_samples)
plot_ordination(
  mature_physeq,
  ordination = ordinate(mature_physeq, method = "PCoA", distance = sm_rel_unifracs$d_1),
  color = "Host_diet"
) + 
  theme(legend.box.background = element_rect(color = NA),
        aspect.ratio = 1) + 
  scale_color_manual(values = c("#196689", "#B6AA0D", "darkred")) + 
  labs(color = "Host Diet") + 
  theme(legend.position = "none")+ 
  geom_point(aes(size = Sample_Sum), alpha = 0.6) + 
  scale_size_continuous(labels = scales::label_log())

```

# Error propagation

```{r, error_prop, fig.width = 4, fig.height = 4}

cell_counts_prep <- 
  mature_physeq %>%
  sample_data %>%
  data.frame %>%
  select(sample, Sample_Sum)

cell_counts <- cell_counts_prep$Sample_Sum

names(cell_counts) = cell_counts_prep$sample

err_01_tables <- vector(mode = "list", length = 50)
err_05_tables <- vector(mode = "list", length = 50)
err_10_tables <- vector(mode = "list", length = 50)
err_25_tables <- vector(mode = "list", length = 50)
err_50_tables <- vector(mode = "list", length = 50)

set.seed(314)

for(i in 1:50){
  
  direction <- rbinom(length(cell_counts), size = 1, prob = 0.5) * 2 - 1
  
  err_01 <- cell_counts + (cell_counts * direction * rnorm(n = length(cell_counts), mean = 0.01, sd = 0.001))
  err_05 <- cell_counts + (cell_counts * direction * rnorm(n = length(cell_counts), mean = 0.05, sd = 0.005))
  err_10 <- cell_counts + (cell_counts * direction * rnorm(n = length(cell_counts), mean = 0.1, sd = 0.01))
  err_25 <- cell_counts + (cell_counts * direction * rnorm(n = length(cell_counts), mean = 0.25, sd = 0.025))
  err_50 <- cell_counts + (cell_counts * direction * rnorm(n = length(cell_counts), mean = 0.5, sd = 0.05))
  
  err_01_table <- generate_rarefied_abs_tables(physeq = mature_physeq, # Physeq with non-normalized counts (do not convert to rel abundance!)
                                                 iterations = 1, # Number of rarefy steps
                                                 rare.depth = min(sample_sums(mature_physeq)), # Depth to rarefy to
                                                 abs.counts = err_01, # NAMED vector of cell counts for each sample
                                                 seed = 314, # Seed determines rarefaction, but not absolute normalization
                                                 return_rel = FALSE)
  
  err_05_table <- generate_rarefied_abs_tables(physeq = mature_physeq, # Physeq with non-normalized counts (do not convert to rel abundance!)
                                                 iterations = 1, # Number of rarefy steps
                                                 rare.depth = min(sample_sums(mature_physeq)), # Depth to rarefy to
                                                 abs.counts = err_05, # NAMED vector of cell counts for each sample
                                                 seed = 314, # Seed determines rarefaction, but not absolute normalization
                                                 return_rel = FALSE)
  
  err_10_table <- generate_rarefied_abs_tables(physeq = mature_physeq, # Physeq with non-normalized counts (do not convert to rel abundance!)
                                                 iterations = 1, # Number of rarefy steps
                                                 rare.depth = min(sample_sums(mature_physeq)), # Depth to rarefy to
                                                 abs.counts = err_10, # NAMED vector of cell counts for each sample
                                                 seed = 314,# Seed determines rarefaction, but not absolute normalization
                                                 return_rel = FALSE)
  
  err_25_table <- generate_rarefied_abs_tables(physeq = mature_physeq, # Physeq with non-normalized counts (do not convert to rel abundance!)
                                                 iterations = 1, # Number of rarefy steps
                                                 rare.depth = min(sample_sums(mature_physeq)), # Depth to rarefy to
                                                 abs.counts = err_25, # NAMED vector of cell counts for each sample
                                                 seed = 314,# Seed determines rarefaction, but not absolute normalization
                                                 return_rel = FALSE)
  
  err_50_table <- generate_rarefied_abs_tables(physeq = mature_physeq, # Physeq with non-normalized counts (do not convert to rel abundance!)
                                                   iterations = 1, # Number of rarefy steps
                                                   rare.depth = min(sample_sums(mature_physeq)), # Depth to rarefy to
                                                   abs.counts = err_50, # NAMED vector of cell counts for each sample
                                                   seed = 314,# Seed determines rarefaction, but not absolute normalization
                                                   return_rel = FALSE)
  err_01_tables[[i]] <- err_01_table[,1,]
  err_05_tables[[i]] <- err_05_table[,1,]
  err_10_tables[[i]] <- err_10_table[,1,]
  err_25_tables[[i]] <- err_25_table[,1,]
  err_50_tables[[i]] <- err_50_table[,1,]
}


norm_table <- generate_rarefied_abs_tables(physeq = mature_physeq, # Physeq with non-normalized counts (do not convert to rel abundance!)
                              iterations = 1, # Number of rarefy steps
                              rare.depth = min(sample_sums(mature_physeq)), # Depth to rarefy to
                              abs.counts = cell_counts, # NAMED vector of cell counts for each sample
                              seed = 314,
                              return_rel = FALSE)[,1,] # Random seed for reproducibility




norm_bray <- vegdist(norm_table, method = "bray")

err_01_brays <- map(err_01_tables, \(x)vegdist(x, method = "bray"))
err_05_brays <- map(err_05_tables, \(x)vegdist(x, method = "bray"))
err_10_brays <- map(err_10_tables, \(x)vegdist(x, method = "bray"))
err_25_brays <- map(err_25_tables, \(x)vegdist(x, method = "bray"))
err_50_brays <- map(err_50_tables, \(x)vegdist(x, method = "bray"))

err_01_diffs_bray <- c(unlist(map(err_01_brays, \(x)x - norm_bray)))
err_05_diffs_bray <- c(unlist(map(err_05_brays, \(x)x - norm_bray)))
err_10_diffs_bray <- c(unlist(map(err_10_brays, \(x)x - norm_bray)))
err_25_diffs_bray <- c(unlist(map(err_25_brays, \(x)x - norm_bray)))
err_50_diffs_bray <- c(unlist(map(err_50_brays, \(x)x - norm_bray)))


err_01_guns <- 
  map(err_01_tables,
    \(x)GUniFrac(otu.tab = x,
                 tree = phy_tree(mature_physeq),
                 alpha = c(0.2,1),
                 normalize_counts = FALSE,
                 verbose = FALSE))

err_05_guns <- 
  map(err_05_tables,
    \(x)GUniFrac(otu.tab = x,
                 tree = phy_tree(mature_physeq),
                 alpha = c(0.2,1),
                 normalize_counts = FALSE,
                 verbose = FALSE))
err_10_guns <- 
  map(err_10_tables,
    \(x)GUniFrac(otu.tab = x,
                 tree = phy_tree(mature_physeq),
                 alpha = c(0.2,1),
                 normalize_counts = FALSE,
                 verbose = FALSE))

err_25_guns <- 
  map(err_25_tables,
    \(x)GUniFrac(otu.tab = x,
                 tree = phy_tree(mature_physeq),
                 alpha = c(0.2,1),
                 normalize_counts = FALSE,
                 verbose = FALSE))
err_50_guns <- 
  map(err_50_tables,
    \(x)GUniFrac(otu.tab = x,
                 tree = phy_tree(mature_physeq),
                 alpha = c(0.2,1),
                 normalize_counts = FALSE,
                 verbose = FALSE))

norm_gun <- 
  GUniFrac(otu.tab = norm_table,
                 tree = phy_tree(mature_physeq),
                 alpha = c(0.2,1),
                 normalize_counts = FALSE,
                 verbose = FALSE)


err_01_diffs_gun1 <- c(unlist(map(err_01_guns, \(x)as.dist(x$unifracs[,,"d_1"]) - as.dist(norm_gun$unifracs[,,"d_1"]))))
err_05_diffs_gun1 <- c(unlist(map(err_05_guns, \(x)as.dist(x$unifracs[,,"d_1"]) - as.dist(norm_gun$unifracs[,,"d_1"]))))
err_10_diffs_gun1 <- c(unlist(map(err_10_guns, \(x)as.dist(x$unifracs[,,"d_1"]) - as.dist(norm_gun$unifracs[,,"d_1"]))))
err_25_diffs_gun1 <- c(unlist(map(err_25_guns, \(x)as.dist(x$unifracs[,,"d_1"]) - as.dist(norm_gun$unifracs[,,"d_1"]))))
err_50_diffs_gun1 <- c(unlist(map(err_50_guns, \(x)as.dist(x$unifracs[,,"d_1"]) - as.dist(norm_gun$unifracs[,,"d_1"]))))

err_01_diffs_gun2 <- c(unlist(map(err_01_guns, \(x)as.dist(x$unifracs[,,"d_0.2"]) - as.dist(norm_gun$unifracs[,,"d_0.2"]))))
err_05_diffs_gun2 <- c(unlist(map(err_05_guns, \(x)as.dist(x$unifracs[,,"d_0.2"]) - as.dist(norm_gun$unifracs[,,"d_0.2"]))))
err_10_diffs_gun2 <- c(unlist(map(err_10_guns, \(x)as.dist(x$unifracs[,,"d_0.2"]) - as.dist(norm_gun$unifracs[,,"d_0.2"]))))
err_25_diffs_gun2 <- c(unlist(map(err_25_guns, \(x)as.dist(x$unifracs[,,"d_0.2"]) - as.dist(norm_gun$unifracs[,,"d_0.2"]))))
err_50_diffs_gun2 <- c(unlist(map(err_50_guns, \(x)as.dist(x$unifracs[,,"d_0.2"]) - as.dist(norm_gun$unifracs[,,"d_0.2"]))))

big_diff_table <- data.frame(Brays = abs(c(err_01_diffs_bray, err_05_diffs_bray, err_10_diffs_bray, err_25_diffs_bray, err_50_diffs_bray)),
           Gun1s = abs(c(err_01_diffs_gun1, err_05_diffs_gun1, err_10_diffs_gun1, err_25_diffs_gun1, err_50_diffs_gun1)),
           Gun2s = abs(c(err_01_diffs_gun2, err_05_diffs_gun2, err_10_diffs_gun2, err_25_diffs_gun2, err_50_diffs_gun2)),
           Error_Perc = rep(c(0.01, 0.05, 0.1, 0.25, 0.5), each = length(err_01_diffs_bray)),
           Actual_Bray = rep(c(norm_bray),250),
           Actual_Gun1 = rep(c(as.dist(norm_gun$unifracs[,,"d_1"])), 250),
           Actual_Gun2 = rep(c(as.dist(norm_gun$unifracs[,,"d_0.2"])), 250),
           Trial = rep(rep(1:50, each = 990), 5))


big_diff_table %>%
  pivot_longer(Brays:Gun2s, names_to = "Metric", values_to = "Difference") %>%
  group_by(Metric, Error_Perc, Trial) %>%
  summarize(mean_diff = mean(Difference)) %>%
  ggplot(aes(x = Error_Perc, y = mean_diff, color = Metric)) +
  stat_summary(geom = "errorbar", width = 0.02,
               fun.min = \(x)mean(x) - sd(x),
               fun.max = \(x)mean(x) + sd(x)) + 
   stat_summary(geom = "point", fun = mean, size = 3) +
  geom_smooth(formula = y~x , method = "lm", linetype = 2, linewidth = 0.5, se = FALSE) + 
  stat_regline_equation(label.y.npc = 0.9) +
  scale_color_manual(values = c("#196689", "#B6AA0D", "darkred")) + 
  theme(legend.position = "none") + 
  labs(x = "Percent Error Added",
       y = "Mean Difference from True")

big_diff_table %>%
  pivot_longer(Brays:Gun2s, names_to = "Metric", values_to = "Difference") %>%
  group_by(Error_Perc, Metric, Trial) %>%
  summarize(max_diff = max(Difference)) %>%
  ggplot(aes(x = Error_Perc, y = max_diff, color = Metric)) +
  stat_summary(geom = "errorbar", width = 0.02,
               fun.min = \(x)mean(x) - sd(x),
               fun.max = \(x)mean(x) + sd(x)) + 
   stat_summary(geom = "point", fun = mean, size = 3) +
  geom_smooth(formula = y~x , method = "lm", linetype = 2, linewidth = 0.5, se = FALSE) + 
  stat_regline_equation(label.y.npc = 0.9) +
  scale_color_manual(values = c("#196689", "#B6AA0D", "darkred")) + 
  theme(legend.position = "none") + 
  labs(x = "Percent Error Added",
       y = "Max Difference from True")
  

```



```{r random-plotting, eval = FALSE}
big_diff_table %>%
  ggplot(aes(x = Actual_Bray, y = Brays)) + 
  geom_point(alpha = 0.01) + 
  facet_wrap(~Error_Perc, scales = "free_y") 

big_diff_table %>%
  ggplot(aes(x = Actual_Gun1, y = Gun1s)) + 
  geom_point(alpha = 0.01) + 
  facet_wrap(~Error_Perc, scales = "free_y") 

big_diff_table %>%
  ggplot(aes(x = Actual_Gun2, y = Gun2s)) + 
  geom_point(alpha = 0.01) + 
  facet_wrap(~Error_Perc, scales = "free_y")

big_diff_table %>%
  filter(Error_Perc == 0.01) %>%
  ggplot(aes(x = Brays, y = Gun1s)) + 
  geom_point(alpha= 0.01) + 
  geom_abline(slope = 1, intercept = 0, linetype = 2, linewidth = 0.2) + 
  theme(legend.position = "none") +
  coord_fixed(xlim = c(0, 0.01), ylim = c(0, 0.01))

big_diff_table %>%
  filter(Error_Perc == 0.1) %>%
  ggplot(aes(x = Brays, y = Gun1s)) + 
  geom_point(alpha= 0.01) + 
  geom_abline(slope = 1, intercept = 0, linetype = 2, linewidth = 0.2) + 
  theme(legend.position = "none") + 
  coord_fixed(xlim = c(0, 0.1), ylim = c(0, 0.1))

big_diff_table %>%
  filter(Error_Perc == 0.5) %>%
  ggplot(aes(x = Brays, y = Gun1s)) + 
  geom_point(alpha= 0.01) + 
  geom_abline(slope = 1, intercept = 0, linetype = 2, linewidth = 0.2) + 
  theme(legend.position = "none") +
  coord_fixed(xlim = c(0, 0.5), ylim = c(0, 0.5))

big_diff_table %>%
  filter(Error_Perc == 0.01) %>%
  ggplot(aes(x = Brays, y = Gun2s)) + 
  geom_point(alpha= 0.01) + 
  geom_abline(slope = 1, intercept = 0, linetype = 2, linewidth = 0.2) + 
  theme(legend.position = "none") +
  coord_fixed(xlim = c(0, 0.01), ylim = c(0, 0.01))

big_diff_table %>%
  filter(Error_Perc == 0.1) %>%
  ggplot(aes(x = Brays, y = Gun2s)) + 
  geom_point(alpha= 0.01) + 
  geom_abline(slope = 1, intercept = 0, linetype = 2, linewidth = 0.2) + 
  theme(legend.position = "none") + 
  coord_fixed(xlim = c(0, 0.1), ylim = c(0, 0.1))

big_diff_table %>%
  filter(Error_Perc == 0.5) %>%
  ggplot(aes(x = Brays, y = Gun2s)) + 
  geom_point(alpha= 0.01) + 
  geom_abline(slope = 1, intercept = 0, linetype = 2, linewidth = 0.2) + 
  theme(legend.position = "none") +
  coord_fixed(xlim = c(0, 0.5), ylim = c(0, 0.5))

big_diff_table %>%
  pivot_longer(Brays:Gun2s, names_to = "Metric", values_to = "Difference") %>%
  ggplot(aes(x = Metric, y = Difference, color = Metric)) + 
  geom_boxplot() + 
  facet_wrap(~Error_Perc, scales = "free_y") + 
  theme(legend.position = "none") + 
  stat_compare_means(method = "wilcox", comparisons = list(c(1,2),c(1,3), c(2,3)))

big_diff_table %>%
  pivot_longer(Brays:Gun2s, names_to = "Metric", values_to = "Difference") %>%
  mutate(log_err = log(Error_Perc),
         log_diff = log(Difference)) %>%
  ggplot(aes(x = log_err, y = log_diff, color = Metric)) +
  stat_summary(geom = "point", fun = mean, size = 2) +
  geom_smooth(method = "lm", linetype = 2, linewidth = 0.5) + 
  stat_regline_equation() 

```
